+++ main/CMakeLists.txt
@@ -3,7 +3,7 @@ project(main)
 add_executable(${PROJECT_NAME} main.cc)
 
 target_link_libraries(${PROJECT_NAME}
-        spmm::serial
+        spmm::spmm
         spmm::datamodel
         spmm::test
         )
