+++ spmm/CMakeLists.txt
@@ -1,9 +1,9 @@
-project(serial)
+project(spmm)
 
 aux_source_directory(src DIR_SRCS)
 
 add_library(${PROJECT_NAME} ${DIR_SRCS})
-add_library(spmm::serial ALIAS ${PROJECT_NAME})
+add_library(spmm::spmm ALIAS ${PROJECT_NAME})
 
 target_include_directories(${PROJECT_NAME}
     PUBLIC ${PROJECT_SOURCE_DIR}/include
