+++ test/CMakeLists.txt
@@ -11,5 +11,5 @@ target_include_directories(${PROJECT_NAME}
 
 target_link_libraries(${PROJECT_NAME}
     spmm::datamodel
-    spmm::serial
+    spmm::spmm
 )
